simulator/flask app/requirements.txt
Flask==3.0.1
Flask-SocketIO==5.4.1
websocket-client==1.8.0